refactor shooter init into client-only factory
